requirements.txt
anyio==4.1.0
appdirs==1.4.4
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cssselect==1.2.0
fake-useragent==1.4.0
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
importlib-metadata==6.8.0
instaloader==4.10.1
lxml==4.9.3
outcome==1.3.0.post0
parse==1.20.0
phantomjs==1.4.1
Pillow==10.1.0
playwright==1.40.0
pycparser==2.21
pyee==11.0.1
pyppeteer==1.0.2
pyquery==2.0.0
PySocks==1.7.1
pyTelegramBotAPI==4.14.0
requests==2.31.0
requests-html==0.10.0
selenium==4.15.2
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
tqdm==4.66.1
trio==0.23.1
trio-websocket==0.11.1
urllib3==1.26.18
w3lib==2.1.2
websockets==10.4
wsproto==1.2.0
zipp==3.17.0
